+++ internal/config/config.go
@@ -9,6 +9,7 @@ import (
 	"reflect"
 	"slices"
 
+	"github.com/jkroepke/openvpn-auth-oauth2/internal/utils"
 	flag "github.com/spf13/pflag"
 )
 
@@ -21,14 +22,19 @@ type Config struct {
 }
 
 type Http struct {
-	Listen               string `koanf:"listen"`
-	CertFile             string `koanf:"cert"`
-	KeyFile              string `koanf:"key"`
-	Tls                  bool   `koanf:"tls"`
-	BaseUrl              string `koanf:"baseurl"`
-	Secret               string `koanf:"secret"`
-	CallbackTemplate     *template.Template
-	CallbackTemplatePath string `koanf:"callback_template_path"`
+	Listen             string             `koanf:"listen"`
+	CertFile           string             `koanf:"cert"`
+	KeyFile            string             `koanf:"key"`
+	Tls                bool               `koanf:"tls"`
+	BaseUrl            *url.URL           `koanf:"baseurl"`
+	Secret             string             `koanf:"secret"`
+	CallbackTemplate   *template.Template `koanf:"callback-template-path"`
+	Check              *HttpCheck         `koanf:"check"`
+	EnableProxyHeaders bool               `koanf:"enable-proxy-headers"`
+}
+
+type HttpCheck struct {
+	IpAddr bool `koanf:"ipaddr"`
 }
 
 type Log struct {
@@ -37,7 +43,7 @@ type Log struct {
 }
 
 type OpenVpn struct {
-	Addr          string         `koanf:"addr"`
+	Addr          *url.URL       `koanf:"addr"`
 	Password      string         `koanf:"password"`
 	Bypass        *OpenVpnBypass `koanf:"bypass"`
 	AuthTokenUser bool           `koanf:"auth-token-user"`
@@ -48,13 +54,14 @@ type OpenVpnBypass struct {
 }
 
 type OAuth2 struct {
-	Issuer    string           `koanf:"issuer"`
-	Provider  string           `koanf:"provider"`
-	Endpoints *OAuth2Endpoints `koanf:"endpoint"`
-	Client    *OAuth2Client    `koanf:"client"`
-	Scopes    []string         `koanf:"scopes"`
-	Pkce      bool             `koanf:"pkce"`
-	Validate  *OAuth2Validate  `koanf:"validate"`
+	Issuer          *url.URL         `koanf:"issuer"`
+	Provider        string           `koanf:"provider"`
+	AuthorizeParams string           `koanf:"authorize-params"`
+	Endpoints       *OAuth2Endpoints `koanf:"endpoint"`
+	Client          *OAuth2Client    `koanf:"client"`
+	Scopes          []string         `koanf:"scopes"`
+	Pkce            bool             `koanf:"pkce"`
+	Validate        *OAuth2Validate  `koanf:"validate"`
 }
 
 type OAuth2Client struct {
@@ -63,9 +70,9 @@ type OAuth2Client struct {
 }
 
 type OAuth2Endpoints struct {
-	Discovery string `koanf:"discovery"`
-	Auth      string `koanf:"auth"`
-	Token     string `koanf:"token"`
+	Discovery *url.URL `koanf:"discovery"`
+	Auth      *url.URL `koanf:"auth"`
+	Token     *url.URL `koanf:"token"`
 }
 
 type OAuth2Validate struct {
@@ -79,7 +86,7 @@ type OAuth2Validate struct {
 func FlagSet() *flag.FlagSet {
 	f := flag.NewFlagSet("openvpn-auth-oauth2", flag.ContinueOnError)
 	f.Usage = func() {
-		_, _ = fmt.Fprintln(os.Stderr, "Usage of openvpn-auth-oauth2:")
+		fmt.Println("Usage of openvpn-auth-oauth2:")
 		f.PrintDefaults()
 		os.Exit(0)
 	}
@@ -93,13 +100,16 @@ func FlagSet() *flag.FlagSet {
 	f.String("http.secret", "", "Cookie secret. (env: CONFIG_HTTP_SECRET)")
 	f.String("http.key", "", "Path to tls server key. (env: CONFIG_HTTP_KEY)")
 	f.String("http.cert", "", "Path to tls server certificate. (env: CONFIG_HTTP_CERT)")
-	f.String("http.callback_template_path", "", "Path to a HTML file which is displayed at the end of the screen. (env: CONFIG_HTTP_CALLBACK_TEMPLATE_PATH)")
+	f.String("http.callback-template-path", "", "Path to a HTML file which is displayed at the end of the screen. (env: CONFIG_HTTP_CALLBACK_TEMPLATE_PATH)")
+	f.Bool("http.check.ipaddr", false, "Check if client IP in http and VPN is equal. (env: CONFIG_HTTP_CHECK_IPADDR)")
+	f.Bool("http.enable-proxy-headers", false, "Use X-Forward-For http header for client ips. (env: CONFIG_HTTP_ENABLE_PROXY_HEADERS)")
 	f.String("openvpn.addr", "unix:///run/openvpn/server.sock", "openvpn management interface addr. Must start with unix:// or tcp:// (env: CONFIG_OPENVPN_ADDR)")
 	f.String("openvpn.password", "", "openvpn management interface password. (env: CONFIG_OPENVPN_PASSWORD)")
 	f.Bool("openvpn.auth-token-user", true, "Define auth-token-user for all sessions. (env: CONFIG_OPENVPN_AUTH_TOKEN_USER)")
 	f.StringSlice("openvpn.bypass.cn", []string{}, "bypass oauth authentication for CNs. (env: CONFIG_OAUTH2_BYPASS_CN)")
 	f.String("oauth2.issuer", "", "oauth2 issuer. (env: CONFIG_OAUTH2_ISSUER)")
-	f.String("oauth2.provider", "oidc", "oauth2 provider. (env: CONFIG_OAUTH2_PROVIDER)")
+	f.String("oauth2.provider", "generic", "oauth2 provider. (env: CONFIG_OAUTH2_PROVIDER)")
+	f.String("oauth2.authorize-params", "", "additional url query parameter to authorize endpoint. (env: CONFIG_OAUTH2_AUTHORIZE_ENDPOINT)")
 	f.String("oauth2.endpoint.discovery", "", "custom oauth2 discovery url. (env: CONFIG_OAUTH2_ENDPOINT_DISCOVERY)")
 	f.String("oauth2.endpoint.auth", "", "custom oauth2 auth endpoint. (env: CONFIG_OAUTH2_ENDPOINT_AUTH)")
 	f.String("oauth2.endpoint.token", "", "custom oauth2 token endpoint. (env: CONFIG_OAUTH2_ENDPOINT_TOKEN)")
@@ -124,29 +134,37 @@ func Validate(conf *Config) error {
 		"log":     conf.Log,
 	} {
 		if reflect.ValueOf(value).IsNil() {
-			return fmt.Errorf("%s is nil", key)
+			return errors.New(utils.StringConcat(key, " is nil"))
 		}
 	}
 
 	for key, value := range map[string]any{
+		"http.check":       conf.Http.Check,
 		"oauth2.client":    conf.Oauth2.Client,
 		"oauth2.endpoints": conf.Oauth2.Endpoints,
 		"oauth2.validate":  conf.Oauth2.Validate,
 		"openvpn.bypass":   conf.OpenVpn.Bypass,
 	} {
 		if reflect.ValueOf(value).IsNil() {
-			return fmt.Errorf("%s is nil", key)
+			return errors.New(utils.StringConcat(key, " is nil"))
 		}
 	}
 
 	for key, value := range map[string]string{
-		"http.baseurl":     conf.Http.BaseUrl,
 		"http.secret":      conf.Http.Secret,
-		"oauth2.issuer":    conf.Oauth2.Issuer,
 		"oauth2.client.id": conf.Oauth2.Client.Id,
 	} {
 		if value == "" {
-			return fmt.Errorf("%s is required", key)
+			return errors.New(utils.StringConcat(key, " is required"))
+		}
+	}
+
+	for key, value := range map[string]*url.URL{
+		"http.baseurl":  conf.Http.BaseUrl,
+		"oauth2.issuer": conf.Oauth2.Issuer,
+	} {
+		if utils.IsUrlEmpty(value) {
+			return errors.New(utils.StringConcat(key, " is required"))
 		}
 	}
 
@@ -154,51 +172,24 @@ func Validate(conf *Config) error {
 		return errors.New("http.secret requires a length of 16, 24 or 32")
 	}
 
-	if uri, err := url.Parse(conf.OpenVpn.Addr); err != nil {
-		return fmt.Errorf("openvpn.addr: invalid URL. error: %s", err)
-	} else if !slices.Contains([]string{"tcp", "unix"}, uri.Scheme) {
+	if !slices.Contains([]string{"tcp", "unix"}, conf.OpenVpn.Addr.Scheme) {
 		return errors.New("openvpn.addr: invalid URL. only tcp://addr or unix://addr scheme supported")
 	}
 
-	if uri, err := url.Parse(conf.Http.BaseUrl); err != nil {
-		return fmt.Errorf("http.baseurl: invalid URL. error: %s", err)
-	} else if uri.Host == "" {
-		return errors.New("http.baseurl: invalid URL. empty hostname")
-	}
-
-	for key, value := range map[string]string{
+	for key, uri := range map[string]*url.URL{
 		"http.baseurl":              conf.Http.BaseUrl,
 		"oauth2.issuer":             conf.Oauth2.Issuer,
 		"oauth2.endpoint.discovery": conf.Oauth2.Endpoints.Discovery,
 		"oauth2.endpoint.token":     conf.Oauth2.Endpoints.Token,
 		"oauth2.endpoint.auth":      conf.Oauth2.Endpoints.Auth,
 	} {
-		if value == "" {
+		if utils.IsUrlEmpty(uri) {
 			continue
 		}
 
-		uri, err := url.Parse(value)
-		if err != nil {
-			return fmt.Errorf("%s: invalid URL. error: %s", key, err)
-		}
-
 		if !slices.Contains([]string{"http", "https"}, uri.Scheme) {
-			return fmt.Errorf("%s: invalid URL. only http:// or https:// scheme supported", key)
+			return errors.New(utils.StringConcat(key, ": invalid URL. only http:// or https:// scheme supported"))
 		}
-	}
-
-	if (conf.Oauth2.Endpoints.Token != "" && conf.Oauth2.Endpoints.Auth == "") ||
-		(conf.Oauth2.Endpoints.Token == "" && conf.Oauth2.Endpoints.Auth != "") {
-		return errors.New("both oauth2.endpoints.tokenUrl and oauth2.endpoints.authUrl are required")
-	}
-
-	if conf.Http.CallbackTemplatePath != "" {
-		tmpl, err := template.New("callback").ParseFiles(conf.Http.CallbackTemplatePath)
-		if err != nil {
-			return fmt.Errorf("http.callbackTemplatePath: invalid template: %s", err)
-		}
-
-		conf.Http.CallbackTemplate = tmpl
 	}
 
 	return nil