+++ internal/oauth2/provider.go
@@ -2,55 +2,52 @@ package oauth2
 
 import (
 	"context"
-	"fmt"
+	"errors"
 	"log/slog"
 	"net/http"
-	"net/url"
 	"strings"
 
 	"github.com/jkroepke/openvpn-auth-oauth2/internal/config"
+	"github.com/jkroepke/openvpn-auth-oauth2/internal/oauth2/providers/generic"
 	"github.com/jkroepke/openvpn-auth-oauth2/internal/oauth2/providers/github"
-	"github.com/jkroepke/openvpn-auth-oauth2/internal/oauth2/providers/oidc"
 	"github.com/jkroepke/openvpn-auth-oauth2/internal/state"
 	"github.com/jkroepke/openvpn-auth-oauth2/internal/types"
+	"github.com/jkroepke/openvpn-auth-oauth2/internal/utils"
 	"github.com/zitadel/oidc/v2/pkg/client/rp"
 	httphelper "github.com/zitadel/oidc/v2/pkg/http"
-	token "github.com/zitadel/oidc/v2/pkg/oidc"
+	"github.com/zitadel/oidc/v2/pkg/oidc"
 	"golang.org/x/oauth2"
 )
 
 type Provider struct {
 	rp.RelyingParty
-	Connector
+	OidcProvider
 }
 
-type Connector interface {
-	CheckUser(ctx context.Context, session *state.State, user *types.UserData, tokens *token.Tokens[*token.IDTokenClaims]) error
-	GetUser(ctx context.Context, tokens *token.Tokens[*token.IDTokenClaims]) (*types.UserData, error)
+type OidcProvider interface {
+	CheckUser(ctx context.Context, session *state.State, user *types.UserData, tokens *oidc.Tokens[*oidc.IDTokenClaims]) error
+	GetEndpoints(conf *config.Config) (*oauth2.Endpoint, error)
+	GetName() string
+	GetUser(ctx context.Context, tokens *oidc.Tokens[*oidc.IDTokenClaims]) (*types.UserData, error)
 }
 
 // NewProvider returns a [rp.RelyingParty] instance
 func NewProvider(logger *slog.Logger, conf *config.Config) (*Provider, error) {
-	baseUrl, err := url.Parse(conf.Http.BaseUrl)
-	if err != nil {
-		return nil, fmt.Errorf("http.baseurl: %v", err)
-	}
-
-	tokenValidator, err := NewTokenValidateProvider(conf)
+	oidcProvider, err := NewOidcProvider(conf)
 	if err != nil {
 		return nil, err
 	}
 
-	redirectURI := fmt.Sprintf("%s/oauth2/callback", strings.TrimSuffix(conf.Http.BaseUrl, "/"))
+	redirectURI := utils.StringConcat(strings.TrimSuffix(conf.Http.BaseUrl.String(), "/"), "/oauth2/callback")
 
 	cookieKey := []byte(conf.Http.Secret)
 	cookieOpt := []httphelper.CookieHandlerOpt{
 		httphelper.WithMaxAge(1800),
-		httphelper.WithPath(fmt.Sprintf("%s/oauth2", strings.TrimSuffix(baseUrl.Path, "/"))),
-		httphelper.WithDomain(baseUrl.Hostname()),
+		httphelper.WithPath(utils.StringConcat(strings.TrimSuffix(conf.Http.BaseUrl.Path, "/"), "/oauth2")),
+		httphelper.WithDomain(conf.Http.BaseUrl.Hostname()),
 	}
 
-	if baseUrl.Scheme == "http" {
+	if conf.Http.BaseUrl.Scheme == "http" {
 		cookieOpt = append(cookieOpt, httphelper.WithUnsecure())
 	}
 
@@ -79,16 +76,21 @@ func NewProvider(logger *slog.Logger, conf *config.Config) (*Provider, error) {
 		options = append(options, rp.WithPKCE(cookieHandler))
 	}
 
-	if conf.Oauth2.Endpoints.Auth == "" && conf.Oauth2.Endpoints.Token == "" {
-		if conf.Oauth2.Endpoints.Discovery != "" {
-			logger.Info(fmt.Sprintf("discover OIDC auto configuration for issuer %s with custom discovery url %s", conf.Oauth2.Issuer, conf.Oauth2.Endpoints.Discovery))
-			options = append(options, rp.WithCustomDiscoveryUrl(conf.Oauth2.Endpoints.Discovery))
+	endpoints, err := oidcProvider.GetEndpoints(conf)
+	if err != nil {
+		return nil, err
+	}
+
+	if endpoints == nil {
+		if !utils.IsUrlEmpty(conf.Oauth2.Endpoints.Discovery) {
+			logger.Info(utils.StringConcat("discover OIDC auto configuration with provider ", oidcProvider.GetName(), "for issuer ", conf.Oauth2.Issuer.String(), "with custom discovery url", conf.Oauth2.Endpoints.Discovery.String()))
+			options = append(options, rp.WithCustomDiscoveryUrl(conf.Oauth2.Endpoints.Discovery.String()))
 		} else {
-			logger.Info(fmt.Sprintf("discover OIDC auto configuration for issuer %s", conf.Oauth2.Issuer))
+			logger.Info(utils.StringConcat("discover OIDC auto configuration with provider ", oidcProvider.GetName(), "for issuer ", conf.Oauth2.Issuer.String()))
 		}
 
 		relayingParty, err := rp.NewRelyingPartyOIDC(
-			conf.Oauth2.Issuer,
+			conf.Oauth2.Issuer.String(),
 			conf.Oauth2.Client.Id,
 			conf.Oauth2.Client.Secret,
 			redirectURI,
@@ -102,21 +104,18 @@ func NewProvider(logger *slog.Logger, conf *config.Config) (*Provider, error) {
 
 		return &Provider{
 			RelyingParty: relayingParty,
-			Connector:    tokenValidator,
+			OidcProvider: oidcProvider,
 		}, nil
 	}
 
-	logger.Info(fmt.Sprintf("manually configure oauth2 provider with endpoints %s and %s", conf.Oauth2.Endpoints.Auth, conf.Oauth2.Endpoints.Token))
+	logger.Info(utils.StringConcat("manually configure oauth2 provider with provider ", oidcProvider.GetName(), " and endpoints ", endpoints.AuthURL, " and ", endpoints.TokenURL))
 
 	rpConfig := &oauth2.Config{
 		ClientID:     conf.Oauth2.Client.Id,
 		ClientSecret: conf.Oauth2.Client.Secret,
 		RedirectURL:  redirectURI,
 		Scopes:       conf.Oauth2.Scopes,
-		Endpoint: oauth2.Endpoint{
-			AuthURL:  conf.Oauth2.Endpoints.Auth,
-			TokenURL: conf.Oauth2.Endpoints.Token,
-		},
+		Endpoint:     *endpoints,
 	}
 
 	relayingParty, err := rp.NewRelyingPartyOAuth(rpConfig, options...)
@@ -126,17 +125,17 @@ func NewProvider(logger *slog.Logger, conf *config.Config) (*Provider, error) {
 
 	return &Provider{
 		RelyingParty: relayingParty,
-		Connector:    tokenValidator,
+		OidcProvider: oidcProvider,
 	}, nil
 }
 
-func NewTokenValidateProvider(conf *config.Config) (Connector, error) {
+func NewOidcProvider(conf *config.Config) (OidcProvider, error) {
 	switch conf.Oauth2.Provider {
-	case "oidc":
-		return oidc.NewProvider(conf), nil
+	case "generic":
+		return generic.NewProvider(conf), nil
 	case "github":
 		return github.NewProvider(conf), nil
 	default:
-		return nil, fmt.Errorf("unknown oauth2 provider: %s", conf.Oauth2.Provider)
+		return nil, errors.New(utils.StringConcat("unknown oauth2 provider: ", conf.Oauth2.Provider))
 	}
 }